requirements.txt
fastapi==0.100.0
passlib==1.7.4
pydantic==1.10.11
python-dotenv==1.0.0
python_jose==3.3.0
Requests==2.31.0
SQLAlchemy==1.4.41
sqlmodel==0.0.8
starlette==0.30.0
